240 ce 4871 add web layer tests for chemical service 23

+++ src/main/java/gov/epa/ccte/api/chemical/domain/ChemicalDetail.java
@@ -8,7 +8,6 @@ import jakarta.validation.constraints.Size;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -53,7 +52,7 @@ public class ChemicalDetail {
     private Double monoisotopicMass;
 
     @Column(name = "percent_assays")
-    private BigDecimal percentAssays;
+    private Double percentAssays;
 
     @Column(name = "pubchem_count")
     private Integer pubchemCount;
